rename Tag to SingleTag
